Added the basic structure for the QPS preprocessor
For now, Queries are tokenized into the defined Lexical Tokens using the Tokenizer class.
Tokens will be represented with the Token class.
The QueryParser class will handle validating the grammar-rules based on the tokens generated. (with future plans to identyify the parameters that should be passed to the PKB based on the actual passed query)