Added evaluator strategies and test cases (#188)
* Add dummy modifies and uses strategies

* Change the logic of the query evaluator where the usage of strategies is independent of the required synonym
So, the strategies will be used regardless of the required synonym but dependent on their existence in the query or not
The required synonym will now be handled inside the strategy itself. So, if the required synonym is a variable, the strategy will return a list of variables etc.

* Add ResultTable that Winson wrote in one of his branches

Add isTruthTable boolean & logic to account for the cases where we joinOnColumns with a suchthat clause that amounts to true like follows(1,2) in which case the non-empty table is returned and no intersection is happening

Add getColumnValues function which takes in a column name and return all the values in the columns. This is to be used in the QueryEvaluator.cpp file to select the required synonym from the table

* Write more test cases to account for cases in the lecture and cartesian product

* Change result from vector of strings to to Pointer of ResultTable
Remove the combineResults as it is substituted with the joinOnColumns function from the ResultTable which handles combining
Add logic to return the column values that match the required synonym in the ResultTable
Add logic where if the column doesn't exist then return all WIP

* Include memory header in follows reader file

* Fix minor temporary issue where you have to return something (need to change later and test)

* Fix minor mistake, it should be setTruthTable not get

* QueryEvaluationStrategy:
Make it return a resultTable

StmtStmtStrategy:
Change the StmtStmtStrategy to take in resultTable instead of string vector
Also add all paramters in processing of first and second parameter. This is needed to determine the column names

FollowsStrategy:
Change FollowsStrategy such that it will populate a resultTable instead of string vector - Need to refractor later
Add case where both parameters are synonyms
Change cases that return all statements such as follows(1,2) to return an empty table with truthTable set to be true

* Change FollowsStrategy such that it will populate a resultTable instead of string vector - Need to refractor later
Add case where both parameters are synonyms
Change cases that return all statements such as follows(1,2) to return an empty table with truthTable set to be true

* Add test cases for follows where both parameters are synonym
Test out new follows and it works - must comment out other strategies as they are still not implemented

* Change only return type of the strategies in function definition - still need to change actual implementation to fill resultTable

* test: Add unit test for FollowsStrategy

* test: Add unit tests for ModifiesStrategy

* test: Add unit tests for ParentStrategy

* test: add extra test cases for ModifiesStrategy

* test: Add unit tests for UsesStrategy

* Implement Pattern with Result Table

* merged with master

* created logic flow for modifiesS

* Moved isBothParamWildcard to evaluation strategy since it applies to all evaluation

* created logics to process both synonyms

* Add logics to process first synonym as param

* Completed logics of Modifies Strategy

* Completed logics of UsesS Strategy

* Integrate ResultTable in ParentStrategy
Add case where both parameters are synonyms
Fix error where the variant was ParentT where it should be Parent*

* Add logic where if the resultTable doesn't have the needed column at all, then we retrieve all entitiesin queryevaluator
Add function in ResultTable.cpp to check if column exists in table

* test: Add unit tests for StmtStmtStrategy

* test: Add unit test skeleton for SuchThatStrategy

* chore: Add documentation and clean up for FollowsStrategy

* chore: Refactor test file names

* Added new method to check for pattern type in result table

* Finished all cases for pattern

* removed CATCH_CONFIG_MAIN in new test cases

* Added differentiation between S and P types for modifies and uses

* feat: Incorporate while and if patterns in header file

* feat: Incorporate If, While pattern in PatternStrategy

* fixed get all entities statement to stmt

* Modifies and Uses bug fixes and tests

* Change some test cases

* change method name for extensibility

* Restore accidentally deleted aigen label

* Changed style as requested

---------

Co-authored-by: MSI\vegit <vegita2009@hotmail.com>
Co-authored-by: TariqBahatheq <82088609+Tariq-droid@users.noreply.github.com>
Co-authored-by: shurvirarora <shurvirarora@gmail.com>
Co-authored-by: Shurvir Arora <63664716+shurvirarora@users.noreply.github.com>